major fixes to predicates datatype warnings also fixed registers are now vectors (again) but with fixed size (maxElements) and there is an attribute validIndex that indicates the number of valid register lanes (similar to HW specs) predicates need further testing

+++ spike/riscv/insns/ss_end_mod_siz_inc.h
@@ -2,7 +2,7 @@ auto &destReg = P.SU.registers[insn.uve_rd()];
 auto sizeReg = insn.uve_mod_size();
 auto dispReg = insn.uve_mod_disp();
 
-int32_t size = READ_REG(sizeReg);
+size_t size = READ_REG(sizeReg);
 int32_t disp = READ_REG(dispReg);
 
 /* For debug, print the indexes of each register as well as their contents
