+++ spike/riscv/insns/ss_sta_ld_d.h
@@ -6,9 +6,9 @@ auto baseReg = insn.uve_conf_base();
 auto sizeReg = insn.uve_conf_size();
 auto strideReg = insn.uve_conf_stride();
 
-reg_t base = READ_REG(baseReg);
-int32_t size = READ_REG(sizeReg);
-int32_t stride = READ_REG(strideReg);
+reg_t  base = READ_REG(baseReg);
+size_t size = READ_REG(sizeReg);
+size_t stride = READ_REG(strideReg);
 
 P.SU.makeStreamRegister<std::uint64_t>(RegisterConfig::Load, streamReg);
 /*operateRegister(P.SU, streamReg, [=](auto& reg) {
